Add uid option to useRequest.
You can pass in a uid that will be used as the object key in the global store and XHR cache, instead of the endpoint. This way, if you have mutliple requests to the same endpoint but with different body values, you can manage if they should be cached separately.